% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{adjust.command}
\alias{adjust.command}
\title{Prepends a ./ to commands on non-Windows systems.}
\usage{
adjust.command(command)
}
\arguments{
\item{command}{the R function or executable representing the problem}
}
\description{
Prepends a ./ to commands on non-Windows systems.
}

