% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{analyze.cart}
\alias{analyze.cart}
\title{Determines the vulernabilities due to deep uncertainties using Classification
and Regression Trees (CART).}
\usage{
analyze.cart(factors, response)
}
\arguments{
\item{factors}{the sampled deeply uncertain parameterizations}

\item{response}{vector of responses whose length equals the number of factors}
}
\description{
Determines the vulernabilities due to deep uncertainties using Classification
and Regression Trees (CART).
}

