% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootstats}
\alias{bootstats}
\title{Internal method copied from the statistics library.}
\usage{
bootstats(b, conf = 0.95, type = "norm")
}
\arguments{
\item{b}{see original documentation}

\item{conf}{see original documentation}

\item{type}{see original documentation}
}
\description{
This function is not exported from the statistics library, so it is copied
here.  This is used to compute the bootstrap confidence intervals for the
Plischke method.
}

