% Generated by roxygen2 (4.0.2): do not edit by hand
\name{borg.optimize.external}
\alias{borg.optimize.external}
\title{Optimize the problem using the Borg standalone executable (borg.exe).}
\usage{
borg.optimize.external(problem, NFE, executable = "borg.exe",
  output = tempfile(), output.frequency = 100, return.output = TRUE,
  verbose = TRUE)
}
\arguments{
\item{problem}{the problem definition}

\item{NFE}{the maximum number of function evaluations}

\item{executable}{the path the the optimization executable}

\item{output}{the location where the runtime output is stored}

\item{output.frequency}{the frequency at which data is output}

\item{return.output}{if \code{TRUE}, this method loads and returns the
contents of the output file}

\item{verbose}{displays additional information for debugging}
}
\description{
This method is used to optimize a problem defined by an external executable.
The Borg MOEA communicates with the external executable using the open API
standardized by the MOEA Framework (\url{http://moeaframework.org}).  See
section 5.2 in the user manual for details of using the API.  Since borg.exe
targets POSIX systems, this method is typically not available on Windows
unless you are running inside Cygwin.  See the Borg MOEA documentation for
instructions on compiling borg.exe.
}
\details{
The Borg MOEA is free and open for non-commercial users.  Source code can
be obtained from \url{http://borgmoea.org}.
}

