% Generated by roxygen2 (4.0.2): do not edit by hand
\name{borg.optimize.function}
\alias{borg.optimize.function}
\title{Optimize a problem using the Borg shared library (borg.dll or libborg.so).}
\usage{
borg.optimize.function(problem, NFE, ...)
}
\arguments{
\item{problem}{the problem definition}

\item{NFE}{the maximum number of function evaluations}

\item{...}{additional arguments for setting algorithm parameters}
}
\description{
This method is used to optimize a problem defined by an R function.  This
method uses R's foreign function interface (FFI) to pass the R function
to the Borg MOEA shared library for optimization.  See the Borg MOEA
documentation for instructions on compiling borg.dll or libborg.so.
}
\details{
The function should either return a vector containing the objectives and
any constraints (e.g., \code{c(o1, o2, o3, c1, c2)}), or a list containing
the objectives and constraints as separate elements
(e.g., \code{list(c(o1, o2, o3), c(c1, c2))}).  All objectives are minimized.
Any non-zero constraint value is considered a constraint violation.

The Borg MOEA is free and open for non-commercial users.  Source code can
be obtained from \url{http://borgmoea.org}.
}

