% Generated by roxygen2 (4.0.2): do not edit by hand
\name{borg.optimize}
\alias{borg.optimize}
\title{Optimize the problem using the borg.exe program.}
\usage{
borg.optimize(problem, NFE, executable = "./borg.exe", output = tempfile(),
  output.frequency = 100, return.output = TRUE, verbose = TRUE)
}
\arguments{
\item{problem}{the problem setup}

\item{NFE}{the maximum number of function evaluations}

\item{executable}{the path the the optimization executable}

\item{output}{the location where the runtime output is stored}

\item{output.frequency}{the frequency at which data is output}

\item{return.output}{if \code{TRUE}, this method loads and returns the
contents of the output file}

\item{verbose}{displays additional information for debugging}
}
\description{
Optimizes the problem.  By default, this method uses the Borg MOEA, which
must first be compiled into an executable on your system.  Since the Borg
MOEA targets POSIX systems, this method is typically not available on
Windows unless you are running inside Cygwin.
}
\details{
The Borg MOEA is free and open for non-commercial users.  Source code can
be obtained from \url{http://borgmoea.org}.
}

