% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.bounds}
\alias{check.bounds}
\title{Returns \code{TRUE} if the decision variables are within bounds.}
\usage{
check.bounds(points, problem)
}
\arguments{
\item{points}{the decision variables}

\item{problem}{the problem definition}
}
\description{
Checks the decision variables to ensure they are within the problem's
lower and upper bounds.
}
\keyword{internal}

