% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.robustness}
\alias{check.robustness}
\title{Computes the robustness metric.}
\usage{
check.robustness(output, problem, method = "default", verbose = FALSE, ...)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{method}{the robustness metric to use (default, variance, constraints,
infogap, or distance)}

\item{verbose}{display additional information}

\item{...}{additional arguments passed to the robustness metric}
}
\description{
Robustness is represented as a scalar value, where values nearer to
positive infinity are considered more robust.  Due to differences in how
each robustness metric computes its value, you should look at relative
differences in values rather than absolute differences.
}

