% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cleanup}
\alias{cleanup}
\title{Removes any sampled point that has NA values.}
\usage{
cleanup(data)
}
\arguments{
\item{data}{the evaluated points}
}
\description{
Generally an NA indicates a problem with the model.  However, this routine
can be used to prune any invalid entries.
}

