% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.uncertainty.models}
\alias{create.uncertainty.models}
\title{Creates models for each uncertainty parameterization.}
\usage{
create.uncertainty.models(problem, factors)
}
\arguments{
\item{problem}{the problem definition from \code{define.problem}}

\item{factors}{the uncertainty parameterizations from \code{sample.lhs}}
}
\description{
Takes the model, which must be defined using an R function, and creates
many instances of the model for each uncertainty parameterization.  The
names of the uncertainty parameters must match the optional, named
arguments to the R function.
}

