% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default.names}
\alias{default.names}
\title{Creates the default column names.}
\usage{
default.names(nvars, nobjs, nconstrs)
}
\arguments{
\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}
}
\description{
Creates default column names of the form \code{VarN}, \code{ObjN}, and
\code{ConstrN}.
}

