% Generated by roxygen2 (4.0.2): do not edit by hand
\name{define.problem}
\alias{define.problem}
\title{Define a new problem formulation.}
\usage{
define.problem(command, nvars, nobjs, nconstrs = 0, bounds = NULL,
  names = NULL, epsilons = NULL, maximize = NULL)
}
\arguments{
\item{command}{the R function or executable representing the problem}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}

\item{bounds}{the lower and upper bounds for each decision variable}

\item{names}{override the column names}

\item{epsilons}{the epsilon values if using Borg to optimize the problem}

\item{maximize}{vector indicating the columns to be maximized}
}
\description{
Constructs a new problem formulation.  The command can either be an R
function or a command line executable.  If using a command line executable,
the program must follow the MOEA Framework external problem protocol,
typically by using the methods in \code{moeaframework.h}.
}
\details{
If using an R function, the function should return a list containing two
vectors.  The first vector stores the objective values and the second stores
the constraint values.
}

