% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plischke.R
\name{deltamim}
\alias{deltamim}
\title{Plischke's delta-moment sensitivity analysis method.}
\usage{
deltamim(x, y, partition.size = min(ceiling(nrow(x)^(2/(7 + tanh((1500 -
  nrow(x))/500)))), 48), quadrature.points = 110, ks.level = 0.95,
  zero.crossing = "on", kd.estimator = "cheap", kd.width = "auto",
  complement = FALSE, plot.enabled = FALSE, plot.cols = min(ncol(x), 4),
  output.trafo = "off", kd.shape = "epanechnikov", ...)
}
\arguments{
\item{x}{the factors}

\item{y}{the response}

\item{partition.size}{the number of partitions}

\item{quadrature.points}{the number of points in the extimated PDFs}

\item{ks.level}{critical value for the KS statistic}

\item{zero.crossing}{detect zero crossings when \code{"on"}}

\item{kd.estimator}{the kernel density estimator to use (cheap, stats,
diffusion, hist)}

\item{kd.width}{the bandwidth for kernel density estimation, or \code{"auto"}}

\item{complement}{compute sensitivities for the complement}

\item{plot.enabled}{if {TRUE}, generate a plot showing the PDFs}

\item{plot.cols}{the number of factors to display in the plot}

\item{output.trafo}{transformation applied to the responses (off, cdf,
normal, interpol, cdf-tight, cdf-loose)}

\item{kd.shape}{kernel shape (normal, triangle, epanechnikov, box, uniform,
biweight, biquadratic) or a function defining the kernel shape}

\item{...}{additional options}
}
\description{
Plischke's delta-moment sensitivity analysis method.
}

