% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.external}
\alias{evaluate.external}
\title{Evaluates a problem representing a command line executable.}
\usage{
evaluate.external(set, problem)
}
\arguments{
\item{set}{the decision variables (inputs) to the problem}

\item{problem}{the problem definition}
}
\description{
Called by the \code{evaluate} method when the problem is a command line
executable.
}
\keyword{internal}

