% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluates the decision variables for a given problem.}
\usage{
evaluate(set, problem)
}
\arguments{
\item{set}{the decision variables (inputs) to the problem}

\item{problem}{the problem definition}
}
\description{
Evaluates the problem using the given decision variables, returning an
object storing the variables, objectives, and constraints.
}

