% Generated by roxygen2 (4.0.1): do not edit by hand
\name{explore}
\alias{explore}
\title{A web-based tool (powered by Shiny) for exploring high-dimensional data sets.}
\usage{
explore(filename, nvars = NULL, nobjs = NULL, nconstrs = 0,
  names = NULL, bounds = NULL, maximize = NULL, order = 1:nobjs,
  visible.variables = FALSE, plot3d.width = "600px",
  plot3d.height = "500px", welcome.panel = NULL, selection.panel = NULL)
}
\arguments{
\item{filename}{the name of the file, a matrix, or a data frame}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}

\item{names}{see \code{mordm.read} for details}

\item{bounds}{see \code{mordm.read} for details}

\item{maximize}{see \code{mordm.read} for details}

\item{order}{ordering of the objectives in the dropdown menus}

\item{visible.variables}{determines if variables are visible by default}

\item{plot3d.width}{the width of the 3D window}

\item{plot3d.height}{the hight of the 3D window}

\item{welcome.panel}{omordm.plotptional panel for displaying a intro message}

\item{selection.panel}{optional panel for displaying info about the selected
point}
}
\description{
If you are providing a custom welcome or selection panel and would like to
display custom resources, use \code{\link{addResourcePath}} to register the
directory containing the resources.
}

