% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{lhsample}
\alias{lhsample}
\title{Generate Latin Hypercube sampled random inputs.}
\usage{
lhsample(nsamples, problem)
}
\arguments{
\item{nsamples}{the number of samples to generate}

\item{problem}{the problem definition}
}
\description{
Generate Latin Hypercube sampled random inputs.
}

