% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.attributes}
\alias{mordm.attributes}
\title{Extracts the common attributes from the time series data.}
\usage{
mordm.attributes(data)
}
\arguments{
\item{data}{the time series data}
}
\description{
Reads the common attributes associated with each entry in the time series
data and returns them in a matrix.
}

