% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.cart}
\alias{mordm.cart}
\title{Determines the vulernabilities due to deep uncertainties using Classification
and Regression Trees (CART).}
\usage{
mordm.cart(factors, response)
}
\arguments{
\item{factors}{the sampled deeply uncertain parameterizations}

\item{response}{vector of responses whose length equals the number of factors}
}
\description{
Determines the vulernabilities due to deep uncertainties using Classification
and Regression Trees (CART).
}

