% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.cbind}
\alias{mordm.cbind}
\title{Adds extra colums to the end of a data set.}
\usage{
mordm.cbind(set, columns)
}
\arguments{
\item{set}{the data set}

\item{columns}{the extra columns}
}
\description{
Adds extra colums to the end of a data set.
}

