% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.colorize}
\alias{mordm.colorize}
\title{Returns a vector of colors to be used when plotting a data set.}
\usage{
mordm.colorize(set, objectives, mark = NULL, palette = NULL, n = 100,
  offset = 0, colors = NULL, clim = NULL, unmarked = "#888888FF",
  alpha = 1, crev = TRUE)
}
\arguments{
\item{set}{the data set}

\item{objectives}{the objectives being plotted (objectives[5] is color)}

\item{mark}{the optional marking rule}

\item{palette}{the color palette to use, either a function that generates the
color palette or the color palette itself}

\item{n}{the number of distinct colors to display (only used if palette is
a function)}

\item{offset}{DEPRECATED}

\item{colors}{user-defined color values}

\item{clim}{range (lower and upper bounds) of color values}

\item{unmarked}{the color value used for unmarked points}

\item{alpha}{the transparency applied to all colors}

\item{crev}{if \code{TRUE}, reverse the color palette}
}
\description{
The color data that is plotted depends on the options given to this function
in the following order:
\itemize{
  \item{colors}{- displays the user-defined color values}
  \item{mark}{- displays the user-defined marking rules, each rule with a separate color}
  \item{objectives}{- use color values stored in the column defined by \code{objective[5]}}
}
}

