% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.correlation}
\alias{mordm.correlation}
\title{Displays the correlations among pairwise factors.}
\usage{
mordm.correlation(data, ht = 0.75, lt = 0.25, all = FALSE,
  objectives = FALSE)
}
\arguments{
\item{data}{the data set to use}

\item{ht}{the threshold for highly correlated pairs}

\item{lt}{the threshold for uncorrelated pairs}

\item{all}{show all correlations}

\item{objectives}{only compute correlations between objectives}
}
\description{
Computes the pairwise correlations between the decision variables and
objectives and prints the formatted results.
}

