% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.defaultnames}
\alias{mordm.defaultnames}
\title{Creates the default column names.}
\usage{
mordm.defaultnames(nvars, nobjs)
}
\arguments{
\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}
}
\description{
Creates default column names of the form \code{VarN} and \code{ObjN}.
}
\keyword{internal}

