% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.differences}
\alias{mordm.differences}
\title{Identifies key similarities/differences between two sets.}
\usage{
mordm.differences(set1, set2, scale = TRUE, decreasing = TRUE,
  splits = 20, n = NULL, all = FALSE)
}
\arguments{
\item{set1}{the first set}

\item{set2}{the second set}

\item{scale}{if \code{TRUE}, scale the plot based on the range of the data}

\item{decreasing}{if \code{TRUE}, order differences in decreasing order}

\item{splits}{the number of bins used by the histogram method}

\item{n}{the number of variables to plot}

\item{all}{plot all variables}
}
\description{
Computes key differences between two data sets.  This method prints formatted
text as well as creates a histogram plot showing key differences.
}

