% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.evaluate.uncertainties}
\alias{mordm.evaluate.uncertainties}
\title{Computes robustness under deep compute.robustness.}
\usage{
mordm.evaluate.uncertainties(samples, satisficing.fcn = NULL,
  factors = NULL, baseline_factors = NULL, custom.fcn = NULL)
}
\arguments{
\item{samples}{the samples generated by \code{mordm.sample.uncertainties}}

\item{satisficing.fcn}{the satisficing function for computing the two
satisficing robustness metrics}

\item{factors}{matrix of the original compute.robustness factors for use by
Satisficing Type II}

\item{baseline_factors}{the baseline factors (under well characterized uncertainty)}

\item{custom.fcn}{custom robustness function}
}
\description{
Computes robustness under deep compute.robustness.
}

