% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.generate.names}
\alias{mordm.generate.names}
\title{Creates the default column names.}
\usage{
mordm.generate.names(nvars, nobjs, nconstrs = 0)
}
\arguments{
\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}
}
\description{
Creates default column names of the form \code{VarN}, \code{ObjN}, and
\code{ConstrN}.
}
\keyword{internal}

