% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.get.set}
\alias{mordm.get.set}
\title{Returns the individual data set entry to be analyzed.}
\usage{
mordm.get.set(data, index = -1)
}
\arguments{
\item{data}{a time series or individual data set}

\item{index}{if \code{data} is a time series, specifies which entry to
return (default is the last entry)}
}
\description{
Determines which data set will be analyzed.  Almost all methods use this
function to convert their \code{data} arguments into a single data set.
The \code{data} argument can be either a time series or an individual
data set.
}

