% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.identify}
\alias{mordm.identify}
\title{Identify and highlight points using the middle mouse button.}
\usage{
mordm.identify(enabled = TRUE, label = FALSE)
}
\arguments{
\item{enabled}{if \code{TRUE}, enables this functionality}

\item{label}{if \code{TRUE}, a label will be added to the 3D scatter plot
identifying the selected point}
}
\description{
Enables a mouse callback for clicking points on the 3D scatter plot and
identifying those points.  If any secondary plots are displayed (e.g.,
parallel coordinates plot or marking plot), the selected point will become
highlighted in those plots.
}

