% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.mark.box}
\alias{mordm.mark.box}
\title{Creates a marking rule from PRIM boxes.}
\usage{
mordm.mark.box(box, mean, mass)
}
\arguments{
\item{box}{the box generated by \code{\link{mordm.prim}}}

\item{mean}{the mean of the box}

\item{mass}{the mass of the box}
}
\description{
PRIM identifies one or more boxes.  This method converts from the PRIM box
representation to a marking.
}

