% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.difference}
\alias{mordm.mark.difference}
\title{Computes the difference of two markings.}
\usage{
mordm.mark.difference(...)
}
\arguments{
\item{...}{the markings}
}
\description{
Markings behave like sets.  A point is contained within the difference if
it is contained in exactly one individual marking.  This is similar to the
exclusive-or operator.
}

