% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.points}
\alias{mordm.mark.points}
\title{Creates a marking rule identifying specific points.}
\usage{
mordm.mark.points(points)
}
\arguments{
\item{points}{one or more rows from the data set considered within the
marking}
}
\description{
Markings allow the user to highlight specific subsets of the data set.  These
marked sets can subsequently be plotted or used in supported calculations.
}

