% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.mark.rule}
\alias{mordm.mark.rule}
\title{Creates a marking rule based on a function.}
\usage{
mordm.mark.rule(condition)
}
\arguments{
\item{condition}{a function of the form \code{f:x -> boolean}, where \code{x}
is a single row from the data set, returning \code{TRUE} if the row is
part of the marking}
}
\description{
Markings allow the user to highlight specific subsets of the data set.  These
marked sets can subsequently be plotted or used in supported calculations.
}

