% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.subtract}
\alias{mordm.mark.subtract}
\title{Computes the subtraction of two markings.}
\usage{
mordm.mark.subtract(rule1, rule2)
}
\arguments{
\item{rule1}{the first marking}

\item{rule2}{the second marking}
}
\description{
Markings behave like sets.  A point is contained within the subtraction if
it is contained in \code{rule1} but not \code{rule2}.
}

