% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.normalize}
\alias{mordm.normalize}
\title{Normalizes the objectives.}
\usage{
mordm.normalize(data, maximize)
}
\arguments{
\item{data}{the data set}

\item{maximize}{vector indicating the columns to be maximized}
}
\description{
By default, the objectives are all minimized.  Maximized objectives are
negated.  This function negates the maximized objectives, returning them to
their original, positive form.
}

