% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.plot.box}
\alias{mordm.plot.box}
\title{Display plot of PRIM boxes.}
\usage{
mordm.plot.box(data, mark, main = "PRIM Box", scale.width = TRUE,
  bar.width = 3, col = NULL, names = NULL, legend = TRUE,
  defaults = NULL)
}
\arguments{
\item{data}{the original data set}

\item{mark}{a list of the PRIM boxes to display}

\item{main}{the plot title}

\item{scale.width}{if \code{TRUE}, reduce the width of the bars as more PRIM
boxes are displayed}

\item{bar.width}{the width of the bars}

\item{col}{vector of bar colors}

\item{names}{names of each PRIM box to display in the legend}

\item{legend}{if \code{TRUE}, renders a legend on the plot}

\item{defaults}{draw horizontal lines to show default values}
}
\description{
Generates a plot showing the bounds of the PRIM boxes.  Currently only works
well with one or two PRIM boxes.
}

