% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.plot.markings}
\alias{mordm.plot.markings}
\title{Display the markings in a box plot (candle stick plot).}
\usage{
mordm.plot.markings(highlight = NULL)
}
\arguments{
\item{highlight}{highlight vector of row indices to be highlighted in the plot}
}
\description{
Creates a box plot showing the range (lower and upper bounds) encompassed by
each marking.
}

