% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.plot.parallel}
\alias{mordm.plot.parallel}
\title{Display a parallel plot of the current data set.}
\usage{
mordm.plot.parallel(highlight = NULL, alpha = 0.4, label.size = 1,
  line.width = 1, selection.scale = 2, show.variables = FALSE)
}
\arguments{
\item{highlight}{vector of row indices to be highlighted in the plot}

\item{alpha}{the transparency value; or \code{NA}}

\item{label.size}{the font size of labels}

\item{line.width}{the width of lines}

\item{selection.scale}{the scale of the selected line}

\item{show.variables}{hide decision variables from the plot if \code{FALSE}}
}
\description{
Creates a parallel axis or parallel coordinates plot of the current data
set.  All display attributes are taken from the current plotting options.
}

