% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.plot}
\alias{mordm.plot}
\title{Sets the current data set and displays a 3D scatter plot.}
\usage{
mordm.plot(data, mark = NULL, index = -1, objectives = NULL,
  stay = TRUE, identify = TRUE, colors = NULL, clim = NULL,
  ideal = FALSE, selection = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, slim = NULL, window = NULL, alpha = 1, tick.size = 1,
  label.size = 1.2, label.line = 1, radius.scale = 1, bg = "white",
  fg = "black", ...)
}
\arguments{
\item{data}{the data set to be displayed (if data is a time series, then the
last entry in the time series is displayed)}

\item{mark}{a list of the markings to be displayed}

\item{index}{if data is a time series, controls which entries to display
(see \code{\link{mordm.getset}} for details)}

\item{objectives}{vector specifying the objectives to be plotted on the
x, y, z, size, and color axes}

\item{stay}{forces the 3D scatter plot to stay on top of other windows}

\item{identify}{if \code{TRUE}, clicking a point with the middle mouse button
will identify and highlight that point}

\item{ideal}{draw a visual indicator of where the ideal point is on the plot}

\item{selection}{draw a visual indicator on the given row indices}

\item{colors}{override the color values}

\item{xlim}{range (lower and upper bounds) for the x axis}

\item{ylim}{range (lower and upper bounds) for the y axis}

\item{zlim}{range (lower and upper bounds) for the z axis}

\item{slim}{range (lower and upper bounds) of size values}

\item{clim}{range (lower and upper bounds) of color values}

\item{window}{the window size (w, h)}

\item{alpha}{vector of transparency values applied to each point}

\item{tick.size}{the size of the tick labels}

\item{label.size}{the size of axis labels}

\item{label.line}{the offset of the labels}

\item{radius.scale}{scaling factor applied to the size of the points}

\item{bg}{background color}

\item{fg}{foreground color}

\item{...}{additional options passed to \code{\link{plot3d}}}
}
\description{
Creates a 3D scatter plot of the data.  As a side effect, this method sets
a global variable identifying the current data set and plotting attributes
to be used by other plotting methods in this package.  This design allows
you to easily create secondary plots that are consistent with the primary
3D scatter plot.
}

