% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.plot2d}
\alias{mordm.plot2d}
\title{Similar to \code{\link{mordm.plot}} except the generated plot is a 2D scatter plot.}
\usage{
mordm.plot2d(data, mark = NULL, index = -1, objectives = NULL,
  alpha = 1, colors = NULL, clim = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, slim = NULL, ...)
}
\arguments{
\item{data}{the data set to be displayed (if data is a time series, then the
last entry in the time series is displayed)}

\item{mark}{a list of the markings to be displayed}

\item{index}{if data is a time series, controls which entries to display
(see \code{\link{mordm.get.set}} for details)}

\item{objectives}{vector specifying the objectives to be plotted on the
x, y, size, and color axes}

\item{alpha}{vector of transparency values applied to each point}

\item{colors}{override the color values}

\item{clim}{range (lower and upper bounds) of color values}

\item{xlim}{range (lower and upper bounds) for the x axis}

\item{ylim}{range (lower and upper bounds) for the y axis}

\item{zlim}{range (lower and upper bounds) for the z axis}

\item{slim}{range (lower and upper bounds) of size values}

\item{...}{additional options passed to \code{\link{plot}}}
}
\description{
Creates a 2D scatter plot of the data.  As a side effect, this method sets
a global variable identifying the current data set and plotting attributes
to be used by other plotting methods in this package.  This design allows
you to easily create secondary plots that are consistent with the primary
3D scatter plot.
}

