% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.plotmark}
\alias{mordm.plotmark}
\title{Display the markings in a box plot (candle stick plot).}
\usage{
mordm.plotmark(highlight = NULL)
}
\arguments{
\item{highlight}{highlight vector of row indices to be highlighted in the plot}
}
\description{
Creates a box plot showing the range (lower and upper bounds) encompassed by
each marking.
}

