% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.plotops}
\alias{mordm.plotops}
\title{Plot operator probabilities.}
\usage{
mordm.plotops(data, time = FALSE, improvements = FALSE, log = FALSE,
  improvement.nfe = 1000, current = NULL)
}
\arguments{
\item{data}{the time series data from a Borg MOEA run}

\item{time}{if \code{TRUE}, the x-axis displays elapsed time; otherwise the
x-axis represents the number of function evaluations (NFE)}

\item{improvements}{if \code{TRUE}, displays a trace of the number of Pareto
improvements}

\item{log}{if \code{TRUE}, plot the x-axis with a log scale}

\item{improvement.nfe}{the window size in NFE when computing the number of
improvements}

\item{current}{draw a line at the current time}
}
\description{
Creates a plot of the Borg MOEA operator probabilities across an entire
time series.
}

