% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.plotpar}
\alias{mordm.plotpar}
\title{Display a parallel plot of the current data set.}
\usage{
mordm.plotpar(highlight = NULL, alpha = 0.4, label.size = 1,
  line.width = 1, selection.scale = 2)
}
\arguments{
\item{highlight}{vector of row indices to be highlighted in the plot}

\item{alpha}{the transparency value; or \code{NA}}

\item{label.size}{the font size of labels}

\item{line.width}{the width of lines}

\item{selection.scale}{the}
}
\description{
Creates a parallel axis or parallel coordinates plot of the current data
set.  All display attributes are taken from the current plotting options.
}

