% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.prim}
\alias{mordm.prim}
\title{Patient rule induction method.}
\usage{
mordm.prim(data, objective, minimize = TRUE, percentages = FALSE,
  expand = TRUE, ...)
}
\arguments{
\item{data}{the data set}

\item{objective}{speecifies the objective index, column name, function, or
marking to use}

\item{minimize}{if \code{TRUE}, flip the threshold direction so that smaller
values are preferred}

\item{percentages}{display percentages in the printout}

\item{expand}{if \code{TRUE}, sets the paste option to 1 to enable expanding
the boxes to fill as much space as possible}

\item{...}{optional arguments passed to \code{\link{prim.box}}}
}
\description{
Performs the patient rule induction method (PRIM) to identify boxes in input
space that correlate with data exceeding a given threshold.
}

