% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.read}
\alias{mordm.read}
\title{Loads the time series data output from an optimization algorithm.}
\usage{
mordm.read(file, nvars, nobjs, nconstrs = 0, bounds = NULL, names = NULL,
  maximize = NULL, digits = NULL)
}
\arguments{
\item{file}{the file name}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}

\item{bounds}{the lower and upper bounds of each decision variable}

\item{names}{override the column names}

\item{maximize}{vector indicating the columns to be maximized}

\item{digits}{number of digits to retain}
}
\description{
Reads the time series data (a list of matrices) from an optimization
algorithm.  The format is defined by the Borg MOEA and MOEA Framework.
}

