% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.read.xls}
\alias{mordm.read.xls}
\title{Loads a data set stored in a XLS or XLSX file.}
\usage{
mordm.read.xls(file, nvars = NULL, nobjs = NULL, bounds = NULL,
  maximize = NULL, names = NULL, ignore = NULL, metadata = NULL, ...)
}
\arguments{
\item{file}{the file name}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{bounds}{the lower and upper bounds of each decision variable}

\item{maximize}{vector indicating the columns to be maximized}

\item{names}{override the column names}

\item{ignore}{columns to remove from the dataset}

\item{metadata}{columns to retain in a metadata attribute}

\item{...}{optional arguments passed to read.xls}
}
\description{
This method is similar in use to \code{\link{mordm.read.csv}}.  Requires
gdata and its dependencies, including a Perl interpreter on the host system.
}

