% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.select.indices}
\alias{mordm.select.indices}
\title{Returns the row indices in the data set that are marked.}
\usage{
mordm.select.indices(set, marking, not = FALSE, or = FALSE)
}
\arguments{
\item{set}{the data set}

\item{marking}{list of markings}

\item{not}{DEPRECATED}

\item{or}{DEPRECATED}
}
\description{
Applies one or more markings to the data set to determine which rows are
contained within the markings.
}

