% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.select}
\alias{mordm.select}
\title{Returns the subset of rows that are marked.}
\usage{
mordm.select(data, marking, index = -1, not = FALSE, or = FALSE)
}
\arguments{
\item{data}{the data set to be displayed (if data is a time series, then the
last entry in the time series is displayed)}

\item{marking}{list of markings}

\item{index}{if data is a time series, controls which entries to display
(see \code{\link{mordm.getset}} for details)}

\item{not}{DEPRECATED}

\item{or}{DEPRECATED}
}
\description{
Applies one or more markings to the data set and returns the subset that
are contained within the markings.
}

