% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.sensitivity}
\alias{mordm.sensitivity}
\title{Computes the sensitivities of the decision variables.}
\usage{
mordm.sensitivity(data, objective, index = -1, all = FALSE, ...)
}
\arguments{
\item{data}{the data set}

\item{objective}{the objective index, column name, function, or marking}

\item{index}{if data is a time series, controls which entries to display
(see \code{\link{mordm.getset}} for details)}

\item{all}{if \code{TRUE}, include all points from all entries in the time
series; otherwise, only the last entry is included}

\item{...}{additional options for Plischke's method}
}
\description{
Using Plischke's delta-moment sensitivity analysis method, this function
computes the sensitivities using the given data.  As such, this method does
not need to evaluate any new data points, it works with the provided data.
}
\details{
If \code{objective} is a marking, then this computes the sensitivities that
cause a point to be included in the marking.  This functionality is still
experimental.
}

