% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.subset}
\alias{mordm.subset}
\title{Returns a subset of a data set.}
\usage{
mordm.subset(set, columns = 1:ncol(set), rows = 1:nrow(set))
}
\arguments{
\item{set}{the data set}

\item{columns}{the columns to retain}

\item{rows}{the rows to retain}
}
\description{
Returns a subset of a data set.
}

