% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.uncertainty.evaluate}
\alias{mordm.uncertainty.evaluate}
\title{Computes robustness under deep uncertainty.}
\usage{
mordm.uncertainty.evaluate(samples, satisficing.fcn = NULL)
}
\arguments{
\item{samples}{the samples generated by \code{mordm.uncertainty.sample}}

\item{satisficing.fcn}{the satisficing function for computing the two
satisficing robustness metrics}
}
\description{
Computes robustness under deep uncertainty.
}

