% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.uncertainty}
\alias{mordm.uncertainty}
\title{This is the old uncertainty function, no longer used.}
\usage{
mordm.uncertainty(data, nsamples, models, sd = 0, base.model = NULL,
  method = "default", verbose = TRUE)
}
\description{
This is the old uncertainty function, no longer used.
}
\keyword{internal}

