% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.uncertainty.sample}
\alias{mordm.uncertainty.sample}
\title{Computes robustness under deep uncertainty.}
\usage{
mordm.uncertainty.sample(data, nsamples, models, sd = 0, verbose = TRUE)
}
\arguments{
\item{data}{the data set}

\item{nsamples}{the number of samples to generate for each point}

\item{models}{the problem formulations created using \code{setup}}

\item{sd}{scalar or vector specifying the standard deviation for each
decision variable}

\item{verbose}{display additional information}
}
\description{
Adds Gaussian noise to the decision variables and resamples the model output.
The samples are distributed across one or more different models for the
problem.  The result from this method should be passed to
\code{mordm.uncertainty.evaluate} to compute the robustness metrics.
}
\details{
If multiple models are provided, it is assumed that all models have the same
inputs and outputs; they would only differ in the internal calculcations
within the model.
}

