% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nsample}
\alias{nsample}
\title{Generate normally distributed random inputs.}
\usage{
nsample(mean, sd, nsamples, problem)
}
\arguments{
\item{mean}{scalar or vector specifying the mean value for each decision
variable}

\item{sd}{scalar or vector specifying the standard deviation for each
decision variable}

\item{nsamples}{the number of samples to generate}

\item{problem}{the problem definition}
}
\description{
Generate normally distributed random inputs.
}

