% Generated by roxygen2 (4.0.1): do not edit by hand
\name{robustness.constraints}
\alias{robustness.constraints}
\title{Robustness metric based on constraint violations.}
\usage{
robustness.constraints(output, problem, weights = NULL, verbose = FALSE,
  original.point = NULL)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{weights}{unused}

\item{verbose}{unused}

\item{original.point}{unused}
}
\description{
Measures the percentage of the sampled points that violate constraints.
}

