% Generated by roxygen2 (4.0.2): do not edit by hand
\name{robustness.default}
\alias{robustness.default}
\title{Default robustness metric.}
\usage{
robustness.default(output, problem, weights = NULL, verbose = FALSE,
  original.point = NULL)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{weights}{the weights assigned to each objective}

\item{verbose}{display additional information}

\item{original.point}{the original point being analyzed}
}
\description{
The default robustness metric that combines variances and constraint
violations.
}

