% Generated by roxygen2 (4.0.1): do not edit by hand
\name{robustness.distance}
\alias{robustness.distance}
\title{Robustness metric based on distance.}
\usage{
robustness.distance(output, problem, weights = NULL, verbose = FALSE,
  original.point = NULL)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{weights}{unused}

\item{verbose}{unused}

\item{original.point}{the original point being analyzed}
}
\description{
Measures the average distance from the original point to the sampled points.
This is slightly different from variance in that variance is not effected
by translational distance.  I.e., two point clouds have the same variance,
but one is offset more.
}

