% Generated by roxygen2 (4.0.2): do not edit by hand
\name{robustness.gap}
\alias{robustness.gap}
\title{Experimental robustness metric based on info gap.}
\usage{
robustness.gap(output, problem, weights = NULL, verbose = FALSE,
  original.point = NULL)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{weights}{unused}

\item{verbose}{unused}

\item{original.point}{the original point being analyzed}
}
\description{
Info gap measures the distance from the original point to the nearest
constraint boundary.  This experimental implementation approximates this
distance by computing the distance based on the sampled points.
}

