% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample.lhs}
\alias{sample.lhs}
\title{Generates uncertainty samples using Latin hypercube sampling.}
\usage{
sample.lhs(nsamples, ...)
}
\arguments{
\item{nsamples}{the desired number of samples}

\item{...}{the named arguments specifying the bounds of each parameter}
}
\description{
Returns a matrix storing uncertainty samples.  The first argument to this
function is the desired number of samples.  The remaining named arguments
specify the parameter name and its bounds.  For example, the following
generates samples for two parameters, a and b:
   \code{factors <- sample.lhs(100, a=c(0, 10), b=c(-1, 1))}
}

