% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sensitivity.levels}
\alias{sensitivity.levels}
\title{Determines number of replicates for sensitivity analysis.}
\usage{
sensitivity.levels(problem, samples, method)
}
\arguments{
\item{problem}{the problem definition}

\item{samples}{the desired number of samples}

\item{method}{the sensitivity analysis method}
}
\description{
Calculates the number of replicates / levels required by the sensitivity
analysis method to produce approximately the given number of samples
}

