% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sensitivity}
\alias{sensitivity}
\title{Standardized interface for sensitivity analysis methods.}
\usage{
sensitivity(problem, objective, samples, method = "fast99", verbose = FALSE,
  plot = FALSE, raw = FALSE, collapse = TRUE, ...)
}
\arguments{
\item{problem}{the problem definition}

\item{objective}{the function, objective index, or objective name whose
sensitivity is being computed}

\item{samples}{the desired number of samples}

\item{method}{string representation of the sensitivity analysis method
(fast99, sobol, sobol2002, sobol2007, sobolEff, soboljansen,
sobolmara, sobolroalhs, morris, prc, src, or plischke)}

\item{verbose}{if \code{TRUE}, print additional information}

\item{plot}{if \code{TRUE}, generate any output plots}

\item{raw}{if \code{TRUE}, return the raw model output; otherwise return the
standardized output}

\item{collapse}{if \code{TRUE}, collapses the list representation of the
variables, objectives, and constraints into a matrix representation}

\item{...}{additional options passed to the sensitivity analysis method}
}
\description{
Attempts to standardize the use of various sensitivity analysis methods.
Supports all of the methods provided by the sensitivity library except for
those using metamodels.
}
\details{
In addition to using the same inputs for each method, the outputs are also
standardized.  For methods computing the first-order indices, the output
contains the sensitivity indices (\code{Si}) and a ranking (\code{rank}).
Methods computing total-order indices, the output contains the total
sensitivity indices (\code{Si.total}) and the ranking (\code{rank.total}).
Where available, the output may also contain confidence intervals
(\code{Ci} and \code{Ci.total}).
}

