% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup}
\alias{setup}
\title{Setup a new problem formulation.}
\usage{
setup(command, nvars, nobjs, nconstrs = 0, bounds = NULL, names = NULL)
}
\arguments{
\item{command}{the R function or executable representing the problem}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}

\item{bounds}{the lower and upper bounds for each decision variable}

\item{names}{override the column names}
}
\description{
Constructs a new problem formulation.  The command can either be an R
function or a command line executable.  If using a command line executable,
the program must follow the MOEA Framework external problem protocol,
typically by using the methods in \code{moeaframework.h}.
}

