% Generated by roxygen2 (4.0.2): do not edit by hand
\name{with.parameters}
\alias{with.parameters}
\title{Assigns the parameters for the given function, returning a new function.}
\usage{
\method{with}{parameters}(fcn, ...)
}
\arguments{
\item{fcn}{the function}

\item{...}{the values for the named parameters}
}
\description{
Uses the functional package to pre-specify or override the named parameters
for a function.
}

