% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{borg.optimize}
\alias{borg.optimize}
\title{Optimize the problem using the Borg MOEA.}
\usage{
borg.optimize(problem, NFE, ...)
}
\arguments{
\item{problem}{the problem definition}

\item{NFE}{the maximum number of function evaluations}

\item{...}{optional parameters passed to the underlying methods}
}
\description{
Optimizes the problem.  By default, this method uses the Borg MOEA, which
must first be compiled into an executable or shared library on your system.
If the problem references an R function, then you must have available the
shared library (borg.dll or libborg.so).  If the problem is an external
program, then you must have available the Borg executable (borg.exe).  See
\code{\link{borg.optimize.function}} and \code{\link{borg.optimize.external}}
for details of each method.
}
\details{
The Borg MOEA is free and open for non-commercial users.  Source code can
be obtained from \url{http://borgmoea.org}.
}

