% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.length}
\alias{check.length}
\title{Ensures the given set contains the correct number of decision variables.}
\usage{
check.length(set, problem)
}
\arguments{
\item{set}{the decision variables (inputs) to the problem}

\item{problem}{the problem definition}
}
\description{
Ensures the given set contains the correct number of decision variables.
}
\keyword{internal}

