% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plischke.R
\name{deltafast}
\alias{deltafast}
\title{Fast version of Plischke's delta-moment sensitivity analysis method.}
\usage{
deltafast(x, y, M = 24)
}
\arguments{
\item{x}{the factors}

\item{y}{the response}

\item{M}{the number of partitions}
}
\description{
Fast version of Plischke's delta-moment sensitivity analysis method.
}

