% Generated by roxygen2 (4.0.1): do not edit by hand
\name{explore}
\alias{explore}
\title{A web-based tool (powered by Shiny) for exploring high-dimensional data sets.}
\usage{
explore(filename, nvars = NULL, nobjs = NULL, nconstrs = 0,
  names = NULL, bounds = NULL, maximize = NULL, order = NULL,
  visible.variables = FALSE, plot3d.width = "600px",
  plot3d.height = "500px", welcome.panel = NULL, selection.panel = NULL,
  ignore = NULL, metadata = NULL)
}
\arguments{
\item{filename}{the name of the file, a matrix, or a data frame}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{nconstrs}{the number of constraints}

\item{names}{see \code{mordm.read} for details}

\item{bounds}{see \code{mordm.read} for details}

\item{maximize}{see \code{mordm.read} for details}

\item{order}{ordering of the objectives in the dropdown menus}

\item{visible.variables}{determines if variables are visible by default}

\item{plot3d.width}{the width of the 3D window}

\item{plot3d.height}{the hight of the 3D window}

\item{welcome.panel}{omordm.plotptional panel for displaying a intro message}

\item{selection.panel}{optional panel for displaying info about the selected
point}

\item{ignore}{columns to remove from the dataset (CSV/Excel only)}

\item{metadata}{columns to retain in a metadata attribute (CSV/Excel only)}
}
\description{
Starts a Shiny server and launches a webpage to display the given dataset.
The Shiny server will remain running even if the web browser is closed.  You
must interrupt the current R function call (press Ctrl-C or select
Session -> Interrupt R in RStudio) to stop the server.
}
\details{
This method currently supports loading CSV files (.csv), Excel files (.xls
or .xlsx), and MOEA runtime files (any other extension).  For CSV and Excel
files, you may optionally specify nvars and/or nobjs.  If unspecified, the
method will assume every column is an objective.  For MOEA runtime files,
nvars and nobjs are mandatory.

For CSV and Excel files, you can optionally ignore certain columns or
indicate a column is metadata.  Ignored columns are completely removed from
the analysis.  Metadata is not shown in the visualizations, but will be
saved in an attribute (\code{attr(data[[i]], "metadata")}).  For example, if
each point has an associated GIF animation, you could treat the column
storing the file path as metadata, allowing you to retrieve and display the
GIF.

If you are providing a custom welcome or selection panel and would like to
display custom resources, use \code{\link{addResourcePath}} to register the
directory containing the resources.
}

