% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.as.data.frame}
\alias{mordm.as.data.frame}
\title{Converts a data set into a data frame.}
\usage{
mordm.as.data.frame(entry)
}
\arguments{
\item{the}{data set to convert}
}
\description{
When loading data into OpenMORDM, any non-numeric columns are converted into
factors and represented internally as integers.  This method reverses that
process to get a data frame storing the original values.
}

