% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.cbind}
\alias{mordm.cbind}
\title{Adds extra colums to the end of a data set.}
\usage{
mordm.cbind(set, columns)
}
\arguments{
\item{set}{the data set}

\item{columns}{the extra columns}
}
\description{
The added columns are treated like objectives.
}

