% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.extract.attributes}
\alias{mordm.extract.attributes}
\title{Extracts the common attributes from the time series data.}
\usage{
mordm.extract.attributes(data)
}
\arguments{
\item{data}{the time series data}
}
\description{
Reads the common attributes associated with each entry in the time series
data and returns them in a matrix.
}

