% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.mark.intersection}
\alias{mordm.mark.intersection}
\title{Computes the intersection of two markings.}
\usage{
mordm.mark.intersection(...)
}
\arguments{
\item{...}{the markings}
}
\description{
Markings behave like sets.  A point is contained within the intersection if
it is contained in all individual markings.
}

