% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.not}
\alias{mordm.mark.not}
\title{Computes the inverse of a marking.}
\usage{
mordm.mark.not(rule)
}
\arguments{
\item{rule}{the original marking}
}
\description{
Markings behave like sets.  A point is contained within the inverse if it is
not contained within the original marking.
}

