% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.selection}
\alias{mordm.mark.selection}
\title{Creates a marking rule identifying the selected points.}
\usage{
mordm.mark.selection()
}
\description{
Allows the user to select a rectangular region in the 3D scatter plot and
returns a marking containing all points within the selected region.
}

