% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.mark.union}
\alias{mordm.mark.union}
\title{Computes the union of two markings.}
\usage{
mordm.mark.union(...)
}
\arguments{
\item{...}{the markings}
}
\description{
Markings behave like sets.  A point is contained in the union if it is
contained within any individual marking.
}

