% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.plot.matrix}
\alias{mordm.plot.matrix}
\title{Plots a scatter matrix.}
\usage{
mordm.plot.matrix(alpha = 0.4, label.size = 1, slim = NULL)
}
\arguments{
\item{alpha}{the transparency value; or \code{NA}}

\item{label.size}{the font size of labels}

\item{slim}{limits on the point size}
}
\description{
Plots a scatter matrix for the current set.
}

