% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.print.box}
\alias{mordm.print.box}
\title{Print descriptive representation of PRIM boxes.}
\usage{
mordm.print.box(data, mark, threshold = 0.01, digits = 3, indent = "")
}
\arguments{
\item{data}{the original data set}

\item{mark}{the PRIM boxes}

\item{threshold}{fuzzy factor when determining if two numbers are equal}

\item{digits}{number of digits to round numbers}

\item{indent}{character string prepended to each line}
}
\description{
Displays the PRIM boxes in a human-readable way.
}

