% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.rbind}
\alias{mordm.rbind}
\title{Adds extra rows to the end of a data set.}
\usage{
mordm.rbind(set, rows)
}
\arguments{
\item{set}{the data set}

\item{rows}{the extra rows}
}
\description{
Adds extra rows to the end of a data set.
}

