% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.read.csv}
\alias{mordm.read.csv}
\title{Loads a data set stored in a CSV file.}
\usage{
mordm.read.csv(file, nvars = NULL, nobjs = NULL, bounds = NULL,
  maximize = NULL, names = NULL, ...)
}
\arguments{
\item{file}{the file name}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{bounds}{the lower and upper bounds of each decision variable}

\item{maximize}{vector indicating the columns to be maximized}

\item{names}{override the column names}

\item{...}{optional arguments passed to read.csv}
}
\description{
This method assumes the first N columns are decision variables, and all other
columns are objectives.  N is determined by \code{ncol(bounds)}.  Unless
overridden, this method sets \code{check.names=FALSE} and
\code{header=TRUE}.
}

