% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.read.matrix}
\alias{mordm.read.matrix}
\title{Loads a data set stored in a matrix or data.frame.}
\usage{
mordm.read.matrix(mat, nvars = NULL, nobjs = NULL, bounds = NULL,
  maximize = NULL, names = NULL)
}
\arguments{
\item{mat}{the matrix or data.frame}

\item{nvars}{the number of decision variables}

\item{nobjs}{the number of objectives}

\item{bounds}{the lower and upper bounds of each decision variable}

\item{maximize}{vector indicating the columns to be maximized}

\item{names}{override the column names}
}
\description{
This method assumes the first N columns are decision variables, and all other
columns are objectives.  N is determined by \code{ncol(bounds)}.
}

