% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.read.matrix}
\alias{mordm.read.matrix}
\title{Loads a data set stored in a matrix.}
\usage{
mordm.read.matrix(mat, bounds = NULL, maximize = NULL)
}
\arguments{
\item{mat}{the matrix}

\item{bounds}{the lower and upper bounds of each decision variable}

\item{maximize}{vector indicating the columns to be maximized}
}
\description{
This method assumes the first N columns are decision variables, and all other
columns are objectives.  N is determined by \code{ncol(bounds)}.
}

