% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.recommend}
\alias{mordm.recommend}
\title{Make recommendations for analyzing the data.}
\usage{
mordm.recommend(data)
}
\arguments{
\item{data}{the data set to be analyzed}
}
\description{
Performs basic checks to ensure the data is formatted correctly.  If any
issues are identified, then it will attempt to provide details on correcting
or dealing with the problem.
}

