% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mordm.uncertainty.evaluate}
\alias{mordm.uncertainty.evaluate}
\title{Computes robustness under deep uncertainty.}
\usage{
mordm.uncertainty.evaluate(samples, satisficing.fcn = NULL, factors = NULL,
  custom.fcn = NULL)
}
\arguments{
\item{samples}{the samples generated by \code{mordm.uncertainty.sample}}

\item{satisficing.fcn}{the satisficing function for computing the two
satisficing robustness metrics}

\item{factors}{matrix of the original uncertainty factors for use by
Satisficing Type II}

\item{custom.fcn}{custom robustness function}
}
\description{
Computes robustness under deep uncertainty.
}

