% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mordm.uncertainty}
\alias{mordm.uncertainty}
\title{Computes robustness under deep uncertainty.}
\usage{
mordm.uncertainty(data, sd, nsamples, models, base.model = NULL,
  method = "default", verbose = TRUE)
}
\arguments{
\item{data}{the data set}

\item{sd}{scalar or vector specifying the standard deviation for each
decision variable}

\item{nsamples}{the number of samples to generate for each point}

\item{models}{the problem formulations created using \code{setup}}

\item{base.model}{if multiple models are provided, specify which model is the
base (default) model}

\item{method}{the robustness metric or a list of metrics to use (see
\code{\link{check.robustness}} for available options)}

\item{verbose}{display additional information}
}
\description{
Adds Gaussian noise to the decision variables and resamples the model output.
The samples are distributed across one or more different models for the
problem. Then computes one or more robustness metrics.
}
\details{
If multiple models are provided, it is assumed that all models have the same
inputs and outputs; they would only differ in the internal calculcations
within the model.
}

