% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OpenMORDM.R
\name{mordm.weight}
\alias{mordm.weight}
\title{Computes a vector of weighted preferences}
\usage{
mordm.weight(data, weights)
}
\arguments{
\item{data}{the data}

\item{weights}{the vector of weights}
}
\description{
Computes a vector of weighted preferences
}

