% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{robustness.variance}
\alias{robustness.variance}
\title{Robustness metric based on variance.}
\usage{
robustness.variance(output, problem, weights = NULL, verbose = FALSE,
  original.point = NULL)
}
\arguments{
\item{output}{the evaluated points}

\item{problem}{the problem definition}

\item{weights}{the weights assigned to each objective}

\item{verbose}{display additional information}

\item{original.point}{the original point being analyzed}
}
\description{
Measures the variance of the sampled points.
}

