% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{usample}
\alias{usample}
\title{Generate uniformly distributed random inputs.}
\usage{
usample(nsamples, problem)
}
\arguments{
\item{nsamples}{the number of samples to generate}

\item{problem}{the problem definition}
}
\description{
Generate uniformly distributed random inputs.
}

